Update test class naming
